consider resolution elements between pixels. Improves min input resolution from around R 10000 to R 1000 resulting in a scan frame generation time of 10 mins from 1h30.
elements that are between pixels have the proportion in each column calculated and split. This is done assuming a vertical sf (no tilt) which seems valid at mid resolution R ~ 1000 (previous approx. was valid at R~10000) but still shows spiking at R~300. This will be due to the fact the tilt means top part of the element is in one column and the bottom part in the other. Improving this calculation could allow lower resolutions still to be used.